Integrated skip verify, gomail and templates
I have switched the mail sending mechanism to gomail to get more power
out of the sending method. Beside that I have added a flag to skip
certificate verification.

On top of that I have integrated the drone templating to make it
possible to overwrite the email templates optionally.

To be more compatible I have created a plaintext message format with the
help of html2text as well.